Split out files and updated to add a relationship query builder and relationship structure.
