+++ src/Dommel/DontMap.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Dommel
+{
+    internal class DontMap
+    {
+    }
+
+}
